+++ Backend-htn-project/data.csv
@@ -1,0 +1,183 @@
+0,0.05,0.04,-1.05,-1.05
+0,0.05,-0.02,-1.35,-1.35
+0,0.1,-0.13,-1.7,-1.7
+0,0.16,-0.2,-2.13,-2.13
+0,0.14,-0.26,-2.66,-2.66
+0,0.18,-0.34,-3.26,-3.26
+0,0.25,-0.42,-3.93,-3.93
+0,0.3,-0.49,-4.68,-4.68
+0,0.26,-0.51,-5.04,-5.04
+0,0.24,-0.47,-4.59,-4.59
+0,0.2,-0.39,-3.79,-3.79
+0,0.17,-0.31,-2.98,-2.98
+0,0.14,-0.25,-2.49,-2.49
+0,0.12,-0.21,-2.1,-2.1
+0,0.1,-0.18,-1.8,-1.8
+0,0.09,-0.16,-1.59,-1.59
+0,0.08,-0.15,-1.47,-1.47
+0,0.08,-0.14,-1.35,-1.35
+0,0.07,-0.12,-1.2,-1.2
+0,0.06,-0.12,-1.12,-1.12
+0,0.06,-0.11,-1.06,-1.06
+0,0.06,-0.11,-1.03,-1.03
+0,0.06,-0.11,-1.01,-1.01
+0,0.06,-0.11,-0.98,-0.98
+0,0.06,-0.11,-0.96,-0.96
+0,0.07,-0.1,-0.94,-0.94
+0,0.08,-0.1,-0.92,-0.92
+0,0.08,-0.1,-0.89,-0.89
+0,0.08,-0.09,-0.86,-0.86
+0,0.08,-0.09,-0.82,-0.82
+0,0.08,-0.08,-0.79,-0.79
+0,0.08,-0.08,-0.77,-0.77
+0,0.08,-0.08,-0.75,-0.75
+0,0.07,-0.07,-0.73,-0.73
+0,0.07,-0.07,-0.71,-0.71
+0,0.07,-0.07,-0.7,-0.7
+0,0.07,-0.06,-0.68,-0.68
+0,0.07,-0.06,-0.66,-0.66
+0,0.06,-0.05,-0.62,-0.62
+0,0.06,-0.05,-0.59,-0.59
+0,0.06,-0.04,-0.58,-0.58
+0,0.06,-0.04,-0.57,-0.57
+0,0.05,-0.03,-0.55,-0.55
+0,0.05,-0.03,-0.54,-0.54
+0,0.05,-0.03,-0.54,-0.54
+0,0.04,-0.03,-0.54,-0.54
+0,0.04,-0.03,-0.55,-0.55
+0,0.04,-0.03,-0.55,-0.55
+0,0.04,-0.03,-0.56,-0.56
+0,0.04,-0.03,-0.56,-0.56
+0,0.04,-0.03,-0.56,-0.56
+0,0.04,-0.02,-0.57,-0.57
+0,0.05,-0.0,-0.56,-0.56
+0,0.07,0.04,-0.45,-0.45
+0,0.08,0.07,-0.37,-0.37
+0,0.07,0.08,-0.33,-0.33
+0,0.07,0.08,-0.32,-0.32
+0,0.07,0.08,-0.32,-0.32
+0,0.08,0.08,-0.34,-0.34
+0,0.08,0.08,-0.35,-0.35
+0,0.08,0.09,-0.37,-0.37
+0,0.09,0.09,-0.39,-0.39
+0,0.09,0.1,-0.42,-0.42
+0,0.1,0.11,-0.44,-0.44
+0,0.1,0.12,-0.45,-0.45
+0,0.1,0.12,-0.46,-0.46
+0,0.1,0.13,-0.47,-0.47
+0,0.1,0.13,-0.48,-0.48
+0,0.1,0.13,-0.49,-0.49
+0,0.1,0.14,-0.5,-0.5
+0,0.1,0.14,-0.51,-0.51
+0,0.1,0.14,-0.51,-0.51
+0,0.1,0.14,-0.51,-0.51
+0,0.1,0.14,-0.5,-0.5
+0,0.09,0.14,-0.49,-0.49
+0,0.1,0.13,-0.47,-0.47
+0,0.1,0.13,-0.46,-0.46
+0,0.1,0.13,-0.45,-0.45
+0,0.1,0.12,-0.44,-0.44
+0,0.1,0.12,-0.42,-0.42
+0,0.1,0.12,-0.41,-0.41
+0,0.09,0.11,-0.4,-0.4
+0,0.09,0.11,-0.38,-0.38
+0,0.09,0.11,-0.37,-0.37
+0,0.09,0.11,-0.36,-0.36
+0,0.09,0.1,-0.35,-0.35
+0,0.09,0.1,-0.35,-0.35
+0,0.08,0.1,-0.34,-0.34
+0,0.08,0.1,-0.33,-0.33
+0,0.08,0.1,-0.33,-0.33
+0,0.07,0.1,-0.32,-0.32
+0,0.03,0.08,-0.33,-0.33
+0,0.02,0.08,-0.31,-0.31
+0,0.01,0.07,-0.3,-0.3
+0,0.01,0.07,-0.3,-0.3
+0,0.01,0.07,-0.3,-0.3
+0,0.01,0.07,-0.3,-0.3
+0,0.01,0.07,-0.31,-0.31
+0,0.01,0.07,-0.31,-0.31
+0,0.01,0.07,-0.32,-0.32
+0,0.01,0.07,-0.32,-0.32
+0,0.01,0.08,-0.33,-0.33
+0,0.01,0.08,-0.34,-0.34
+0,0.01,0.08,-0.34,-0.34
+0,0.01,0.08,-0.35,-0.35
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.08,-0.36,-0.36
+0,0.02,0.09,-0.35,-0.35
+0,0.03,0.09,-0.35,-0.35
+0,0.03,0.09,-0.35,-0.35
+0,0.03,0.09,-0.35,-0.35
+0,0.03,0.09,-0.34,-0.34
+0,0.03,0.09,-0.34,-0.34
+0,0.03,0.09,-0.34,-0.34
+0,0.03,0.09,-0.33,-0.33
+0,0.02,0.08,-0.33,-0.33
+0,0.02,0.08,-0.32,-0.32
+0,0.02,0.08,-0.31,-0.31
+0,0.02,0.08,-0.31,-0.31
+0,0.02,0.08,-0.31,-0.31
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.02,0.08,-0.3,-0.3
+0,0.03,0.08,-0.29,-0.29
+0,0.03,0.08,-0.28,-0.28
+0,0.04,0.08,-0.28,-0.28
+0,0.04,0.08,-0.27,-0.27
+0,0.04,0.08,-0.27,-0.27
+0,0.04,0.08,-0.27,-0.27
+0,0.04,0.08,-0.28,-0.28
+0,0.04,0.08,-0.28,-0.28
+0,0.04,0.08,-0.28,-0.28
+0,0.04,0.08,-0.29,-0.29
+0,0.04,0.08,-0.29,-0.29
+0,0.04,0.08,-0.3,-0.3
+0,0.04,0.08,-0.3,-0.3
+0,0.05,0.08,-0.31,-0.31
+0,0.05,0.09,-0.32,-0.32
+0,0.05,0.09,-0.32,-0.32
+0,0.04,0.09,-0.33,-0.33
+0,0.04,0.09,-0.33,-0.33
+0,0.04,0.09,-0.33,-0.33
+0,0.04,0.09,-0.34,-0.34
+0,0.04,0.09,-0.35,-0.35
+0,0.04,0.09,-0.35,-0.35
+0,0.04,0.09,-0.36,-0.36
+0,0.04,0.1,-0.37,-0.37
+0,0.04,0.1,-0.38,-0.38
+0,0.04,0.1,-0.39,-0.39
+0,0.04,0.1,-0.4,-0.4
+0,0.04,0.1,-0.4,-0.4
+0,0.04,0.1,-0.41,-0.41
+0,0.04,0.1,-0.41,-0.41
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.43,-0.43
+0,0.04,0.1,-0.43,-0.43
+0,0.04,0.1,-0.43,-0.43
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.42,-0.42
+0,0.04,0.1,-0.41,-0.41
+0,0.04,0.1,-0.4,-0.4
+0,0.03,0.09,-0.4,-0.4
+0,0.03,0.09,-0.39,-0.39
+0,0.03,0.09,-0.39,-0.39
+0,0.03,0.09,-0.38,-0.38
+0,0.03,0.08,-0.37,-0.37
+0,0.04,0.03,-0.35,-0.35
+0,0.03,-0.01,-0.3,-0.3
+0,0.03,-0.02,-0.27,-0.27
+0,0.03,-0.02,-0.26,-0.26
