+++ Backend-htn-project/master.cmd
@@ -1,0 +1,1 @@
+python app.py & populate_data.py
