+++ Frontend-htn-project/src/App.jsx
@@ -8,7 +8,6 @@ function App() {
 
   return (
     <div style={{display: "flex"}}>
-      <livefeed></livefeed>
     </div>
   )
 }
